Change PlayPause button to JToggleButton
PlayPause button is now used as the source of truth for whether the
simulation is running or not.

Fix input verification of mapSizeField to only allow positive integer
values greater than zero. This correctly updates the state of the
PlayPause and Reset buttons.

Pressing Reset is the only way to generate a new map if the map size has
been changed. Pressing PlayPause will never generate a new map, except
in the case where there is no initial map.